+++ terraform/infra/modules/app/s3.tf
@@ -4,20 +4,22 @@ resource "aws_s3_object" "website_config" {
   key          = "/config.js"
   content_type = "text/javascript"
   content = templatefile("${path.module}/templates/config.js", {
-    pool_id        = var.user_pool_id,
-    pool_client_id = var.user_pool_client_id,
-    region         = var.region,
-    invoke_url     = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
-    domains        = var.domains,
-    control_domain = var.control_domain
+    pool_id             = var.user_pool_id,
+    pool_client_id      = var.user_pool_client_id,
+    region              = var.region,
+    invoke_url          = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
+    domains             = var.domains,
+    control_domain      = var.control_domain,
+    invitation_required = var.invitation_required
   })
   etag = md5(templatefile("${path.module}/templates/config.js", {
-    pool_id        = var.user_pool_id,
-    pool_client_id = var.user_pool_client_id,
-    region         = var.region,
-    invoke_url     = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
-    domains        = var.domains,
-    control_domain = var.control_domain
+    pool_id             = var.user_pool_id,
+    pool_client_id      = var.user_pool_client_id,
+    region              = var.region,
+    invoke_url          = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
+    domains             = var.domains,
+    control_domain      = var.control_domain,
+    invitation_required = var.invitation_required
     })
   )
 }
@@ -32,20 +34,22 @@ resource "aws_s3_object" "website_config_json" {
   key          = "/config.json"
   content_type = "application/json"
   content = templatefile("${path.module}/templates/config.js", {
-    pool_id        = var.user_pool_id,
-    pool_client_id = var.user_pool_client_id,
-    region         = var.region,
-    invoke_url     = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
-    domains        = var.domains,
-    control_domain = var.control_domain
+    pool_id             = var.user_pool_id,
+    pool_client_id      = var.user_pool_client_id,
+    region              = var.region,
+    invoke_url          = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
+    domains             = var.domains,
+    control_domain      = var.control_domain,
+    invitation_required = var.invitation_required
   })
   etag = md5(templatefile("${path.module}/templates/config.js", {
-    pool_id        = var.user_pool_id,
-    pool_client_id = var.user_pool_client_id,
-    region         = var.region,
-    invoke_url     = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
-    domains        = var.domains,
-    control_domain = var.control_domain
+    pool_id             = var.user_pool_id,
+    pool_client_id      = var.user_pool_client_id,
+    region              = var.region,
+    invoke_url          = "https://${aws_api_gateway_rest_api.gateway.id}.execute-api.${var.region}.amazonaws.com/${var.stage_name}",
+    domains             = var.domains,
+    control_domain      = var.control_domain,
+    invitation_required = var.invitation_required
     })
   )
 }