+++ terraform/infra/modules/app/templates/config.js
@@ -2,6 +2,7 @@
   "control_domain": "${control_domain}",
   "domains": ${jsonencode(domains)},
   "invokeUrl": "${invoke_url}",
+  "invitation_required": ${jsonencode(invitation_required)},
   "cognitoConfig": {
     "region": "${region}",
     "poolData": {